Pushing up Sim Params

+++ src/main/java/cse308/Simulation/ObjectiveFuncEvaluator.java
@@ -19,7 +19,7 @@ public class ObjectiveFuncEvaluator {
         return 0;
     }    
     
-    public static float evaluateObjective(FunctionWeights w, Map m){
+    public static double evaluateObjective(FunctionWeights w, Map m){
        return w.w_compactness*calcCompactness(m)+w.w_partisan_fairness*calcPartisanFairness(m)+w.w_population_equality*calcPopulationEquality(m);
     }
 }
